school-console-app/src/main/java/com/foxminded/chendev/schoolconsoleapp/service/impl/CourseServiceImpl.java
package com.foxminded.chendev.schoolconsoleapp.service.impl;

import com.foxminded.chendev.schoolconsoleapp.exception.DataBaseRuntimeException;
import com.foxminded.chendev.schoolconsoleapp.repository.CourseRepository;
import com.foxminded.chendev.schoolconsoleapp.entity.Course;
import com.foxminded.chendev.schoolconsoleapp.service.CourseService;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CourseServiceImpl implements CourseService {

    private final CourseRepository courseRepository;

    public CourseServiceImpl(CourseRepository courseRepository) {

        this.courseRepository = courseRepository;
    }

    @Override
    public List<Course> findAllCourses() {
        try {
            return courseRepository.findAll();
        } catch (Exception e) {
            throw new DataBaseRuntimeException(e);
        }
    }
}
